restructured the future api calls. Check commit for more info
before this we would have needed to call the data in every component on their own now we can call them just once from the page and set the data to the components. Charts don't use this yet because I will have to see how it works because the data viewed can be changed so it is a todo.